+++ src/dolphin/io.py
@@ -18,7 +18,7 @@ from pyproj import CRS
 
 from dolphin._background import _DEFAULT_TIMEOUT, BackgroundReader, BackgroundWriter
 from dolphin._log import get_log
-from dolphin._types import Filename
+from dolphin._types import Bbox, Filename
 from dolphin.utils import gdal_to_numpy_type, numpy_to_gdal_type, progress
 
 gdal.UseExceptions()
@@ -56,7 +56,7 @@ def load_gdal(
     filename: Filename,
     *,
     band: Optional[int] = None,
-    subsample_factor: int = 1,
+    subsample_factor: Union[int, tuple[int, int]] = 1,
     rows: Optional[slice] = None,
     cols: Optional[slice] = None,
     masked: bool = False,
@@ -69,7 +69,7 @@ def load_gdal(
         Path to the file to load.
     band : int, optional
         Band to load. If None, load all bands as 3D array.
-    subsample_factor : int, optional
+    subsample_factor : int or tuple[int, int], optional
         Subsample the data by this factor. Default is 1 (no subsampling).
         Uses nearest neighbor resampling.
     rows : slice, optional
@@ -91,6 +91,9 @@ def load_gdal(
     # Make an output object of the right size
     dt = gdal_to_numpy_type(ds.GetRasterBand(1).DataType)
 
+    if isinstance(subsample_factor, int):
+        subsample_factor = (subsample_factor, subsample_factor)
+
     if rows is not None and cols is not None:
         xoff, yoff = cols.start, rows.start
         row_stop = min(rows.stop, nrows)
@@ -101,11 +104,17 @@ def load_gdal(
                 f"Invalid row/col slices: {rows}, {cols} for file {filename} of size"
                 f" {nrows}x{ncols}"
             )
-        nrows_out, ncols_out = ysize // subsample_factor, xsize // subsample_factor
+        nrows_out, ncols_out = (
+            ysize // subsample_factor[0],
+            xsize // subsample_factor[1],
+        )
     else:
         xoff, yoff = 0, 0
         xsize, ysize = ncols, nrows
-        nrows_out, ncols_out = nrows // subsample_factor, ncols // subsample_factor
+        nrows_out, ncols_out = (
+            nrows // subsample_factor[0],
+            ncols // subsample_factor[1],
+        )
     # Read the data, and decimate if specified
     resamp = gdal.GRA_NearestNeighbour
     if band is None:
@@ -246,8 +255,8 @@ def get_raster_gt(filename: Filename) -> list[float]:
 
     Returns
     -------
-    tuple[float, float, float, float, float, float]
-        Geotransform.
+    List[float]
+        6 floats representing a GDAL Geotransform.
     """
     ds = gdal.Open(fspath(filename))
     gt = ds.GetGeoTransform()
@@ -292,7 +301,7 @@ def get_raster_driver(filename: Filename) -> str:
 
 def get_raster_bounds(
     filename: Optional[Filename] = None, ds: Optional[gdal.Dataset] = None
-) -> tuple[float, float, float, float]:
+) -> Bbox:
     """Get the (left, bottom, right, top) bounds of the image."""
     if ds is None:
         if filename is None:
@@ -638,6 +647,7 @@ class EagerLoader(BackgroundReader):
         nodata_mask: Optional[ArrayLike] = None,
         queue_size: int = 1,
         timeout: float = _DEFAULT_TIMEOUT,
+        show_progress: bool = True,
     ):
         super().__init__(nq=queue_size, timeout=timeout, name="EagerLoader")
         self.filename = filename
@@ -656,6 +666,7 @@ class EagerLoader(BackgroundReader):
         self._nodata_mask = nodata_mask
         self._block_shape = block_shape
         self._nodata = get_raster_nodata(filename)
+        self._show_progress = show_progress
         if self._nodata is None:
             self._nodata = np.nan
 
@@ -681,7 +692,7 @@ class EagerLoader(BackgroundReader):
 
         s_iter = range(len(queued_slices))
         desc = f"Processing {self._block_shape} sized blocks..."
-        with progress() as p:
+        with progress(dummy=not self._show_progress) as p:
             for _ in p.track(s_iter, description=desc):
                 cur_block, (rows, cols) = self.get_data()
                 logger.debug(f"got data for {rows, cols}: {cur_block.shape}")